server/packages/scaffolding/pyproject.toml
[project]
name = "zenve-scaffolding"
version = "0.1.0"
description = "zenve template scaffolding and preset management"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "zenve-config",
    "zenve-models",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/zenve_scaffolding"]

[tool.uv.sources]
zenve-config = { workspace = true }
zenve-models = { workspace = true }
